requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
chromadb==0.4.18
sentence-transformers==2.2.2
google-generativeai==0.3.2
pandas>=2.0.0
numpy>=1.24.0
python-dotenv==1.0.0
requests>=2.31.0
tqdm>=4.65.0
gunicorn>=21.0.0
